Renamed component to MultiText
