+++ src/MultiText/MultiText.js
