+++ src/index.js
@@ -5,7 +5,7 @@ export { default as Copyright } from "./Copyright";
 export { default as Color } from "./Color";
 export { default as FontPicker } from "./FontPicker";
 export { default as FontStyle } from "./FontStyle";
-export { default as LabelSetter } from "./LabelSetter";
+export { default as MultiText } from "./MultiText";
 export { default as Loading } from "./Loading";
 export { default as LoginForm } from "./LoginForm";
 export { default as NoLicense } from "./NoLicense";
